requirements.txt
SpeechRecognition
openai-whisper
pyautogui
sounddevice
pyaudio
comtypes
pyautogui
opencv-python